% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{avail}
\alias{avail}
\title{Availability.}
\usage{
avail(unavailTime, totalTime)
}
\arguments{
\item{unavailTime}{Unavailable Time. A numeric value representing the unavailable time or a
numeric vector of unavailable times.}

\item{totalTime}{Total Time. A numeric value representing the total time for a given period
or a numeric vector of time periods}
}
\value{
The function returns a numeric value representing the availability for a given period.
}
\description{
Availability.
}
\examples{
unavail <- 100
total <- 1000
avail(unavail, total)
}
