% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mttf.R
\name{mtbf}
\alias{mtbf}
\title{Mean Time Between Failures (MTBF).}
\usage{
mtbf(failures, totalTime)
}
\arguments{
\item{failures}{Failures. A numeric value representing the number of failures for a given period.}

\item{totalTime}{Total Time. A numeric value representing the total time for a given period
or a numeric vector of time periods}
}
\value{
The function returns a numeric value representing the MTBF for a given period.
}
\description{
Mean Time Between Failures (MTBF).
}
\examples{
fail <- 5
total <- 1000
mtbf(fail, total)
}
