% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtbf.R
\name{mttf}
\alias{mttf}
\title{Mean Time To Failure (MTTF).}
\usage{
mttf(failures, totalTime)
}
\arguments{
\item{failures}{Failures. A numeric value representing the number of failures for a given period.}

\item{totalTime}{Total Time. A numeric value representing the total time for a given period
or a numeric vector of time periods}
}
\value{
The function returns a numeric value representing the MTTF for a given period.
}
\description{
Mean Time To Failure (MTTF).
}
\examples{
fail <- 5
total <- 1000
mttf(fail, total)
}
