% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{rel}
\alias{rel}
\title{Reliability.}
\usage{
rel(outageTime, totalTime)
}
\arguments{
\item{outageTime}{Forced Outage Time. A numeric value representing the forced outage time or a
numeric vector of outage times.}

\item{totalTime}{Total Time. A numeric value representing the total time for a given period
or a numeric vector of time periods}
}
\value{
The function returns a numeric value representing the reliability for a given period.
}
\description{
Reliability.
}
\examples{
outage <- 100
total <- 1000
rel(outage, total)
}
